backend/requirements.txt
# FastAPI & ASGI Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.1
pydantic[email]==2.5.0
pydantic-settings==2.1.0

# OAuth
authlib==1.2.1
httpx==0.25.1

# Background Tasks
celery==5.3.4
redis==5.0.1

# Web Scraping
beautifulsoup4==4.12.2
requests==2.31.0
lxml==4.9.3

# Google Drive API
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-api-python-client==2.108.0

# GitHub API
PyGithub==2.1.1

# Markdown Processing
markdown==3.5.1
html2text==2020.1.16

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.1
